+++ app/src/main/java/com/skgtecnologia/sisem/data/auth/remote/AuthRemoteDataSource.kt
@@ -5,7 +5,7 @@ import com.skgtecnologia.sisem.data.auth.remote.model.AuthenticateBody
 import com.skgtecnologia.sisem.data.auth.remote.model.mapToDomain
 import com.skgtecnologia.sisem.data.remote.extensions.apiCall
 import com.skgtecnologia.sisem.domain.auth.model.AccessTokenModel
-import com.skgtecnologia.sisem.domain.model.error.ErrorModelFactory
+import com.skgtecnologia.sisem.domain.model.banner.ErrorModelFactory
 import javax.inject.Inject
 
 class AuthRemoteDataSource @Inject constructor(
