+++ app/src/main/java/com/skgtecnologia/sisem/data/auth/remote/model/AccessTokenResponse.kt
@@ -1,7 +1,7 @@
 package com.skgtecnologia.sisem.data.auth.remote.model
 
-import com.skgtecnologia.sisem.data.remote.model.error.ErrorResponse
-import com.skgtecnologia.sisem.data.remote.model.error.mapToDomain
+import com.skgtecnologia.sisem.data.remote.model.banner.BannerResponse
+import com.skgtecnologia.sisem.data.remote.model.banner.mapToDomain
 import com.skgtecnologia.sisem.domain.auth.model.AccessTokenModel
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
@@ -19,7 +19,7 @@ data class AccessTokenResponse(
     @Json(name = "is_admin") val isAdmin: Boolean,
     @Json(name = "user_id") val userId: Int,
     @Json(name = "name_user") val nameUser: String,
-    @Json(name = "warning") val warning: ErrorResponse?,
+    @Json(name = "warning") val warning: BannerResponse?,
     @Json(name = "doc_type") val docType: String,
     @Json(name = "document") val document: String
 )
