+++ app/src/main/java/com/skgtecnologia/sisem/data/deviceauth/remote/DeviceAuthRemoteDataSource.kt
@@ -8,7 +8,7 @@ import com.skgtecnologia.sisem.data.remote.model.screen.Params
 import com.skgtecnologia.sisem.data.remote.model.screen.ScreenBody
 import com.skgtecnologia.sisem.data.remote.model.screen.mapToDomain
 import com.skgtecnologia.sisem.domain.deviceauth.model.AssociateDeviceModel
-import com.skgtecnologia.sisem.domain.model.error.ErrorModelFactory
+import com.skgtecnologia.sisem.domain.model.banner.ErrorModelFactory
 import com.skgtecnologia.sisem.domain.model.screen.ScreenModel
 import javax.inject.Inject
