+++ app/src/main/java/com/skgtecnologia/sisem/data/operation/remote/OperationRemoteDataSource.kt
@@ -4,7 +4,7 @@ import com.skgtecnologia.sisem.commons.extensions.mapResult
 import com.skgtecnologia.sisem.data.operation.remote.model.mapToDomain
 import com.skgtecnologia.sisem.data.remote.extensions.apiCall
 import com.skgtecnologia.sisem.domain.authcards.model.OperationModel
-import com.skgtecnologia.sisem.domain.model.error.ErrorModelFactory
+import com.skgtecnologia.sisem.domain.model.banner.ErrorModelFactory
 import javax.inject.Inject
 
 class OperationRemoteDataSource @Inject constructor(
