+++ app/src/main/java/com/skgtecnologia/sisem/ui/deviceauth/DeviceAuthScreen.kt
@@ -18,7 +18,7 @@ import com.valkiria.uicomponents.action.DeviceAuthUiAction.DeviceAuthCodeInput
 import com.valkiria.uicomponents.action.FooterUiAction
 import com.valkiria.uicomponents.action.GenericUiAction
 import com.valkiria.uicomponents.action.UiAction
-import com.valkiria.uicomponents.components.errorbanner.OnErrorHandler
+import com.valkiria.uicomponents.components.banner.OnErrorHandler
 import com.valkiria.uicomponents.components.loader.OnLoadingHandler
 import kotlinx.coroutines.launch
 import timber.log.Timber
@@ -87,7 +87,7 @@ fun DeviceAuthScreen(
         }
     }
 
-    OnErrorHandler(errorModel = uiState.disassociateInfoModel) {
+    OnErrorHandler(uiModel = uiState.disassociateInfoModel) {
         handleFooterAction(
             uiAction = it,
             viewModel = viewModel,
@@ -95,7 +95,7 @@ fun DeviceAuthScreen(
         )
     }
 
-    OnErrorHandler(errorModel = uiState.errorModel) {
+    OnErrorHandler(uiModel = uiState.errorModel) {
         viewModel.handleShownError()
     }
