+++ app/src/main/java/com/skgtecnologia/sisem/ui/report/AddReportRoleViewModel.kt
@@ -6,7 +6,7 @@ import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.skgtecnologia.sisem.commons.resources.AndroidIdProvider
-import com.skgtecnologia.sisem.domain.model.error.mapToUi
+import com.skgtecnologia.sisem.domain.model.banner.mapToUi
 import com.skgtecnologia.sisem.domain.report.usecases.GetAddReportRoleScreen
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
